MATX-198 - Extracted code common between MaterialXSurfaceOverride and MaterialXTextureOverride to MaterialXShadingNodeImpl.
